+++ projeto/calendar_generator_engine/res/examinationSolverConfig.xml
@@ -2,9 +2,7 @@
 <solver>
   <!--<environmentMode>FAST_ASSERT</environmentMode>-->
   <solutionClass>examination.domain.Examination</solutionClass>
-  <entityClass>examination.domain.Exam</entityClass>
-  <entityClass>examination.domain.LeadingExam</entityClass>
-  <entityClass>examination.domain.FollowingExam</entityClass>
+  <entityClass>examination.domain.RoomPeriod</entityClass>
 
   <scoreDirectorFactory>
     <scoreDefinitionType>HARD_SOFT</scoreDefinitionType>
@@ -13,76 +11,7 @@
   </scoreDirectorFactory>
 
   <termination>
-    <!--
-      Official benchmark secondsSpentLimit allowed on:
-      - ge0ffrey's main pc: 429
-      - ge0ffrey's old pc: 1152
-    -->
-    <!--<secondsSpentLimit>429</secondsSpentLimit>-->
     <secondsSpentLimit>5</secondsSpentLimit>
     <!--<bestScoreLimit>-0hard/-999999soft</bestScoreLimit>-->
   </termination>
-  <constructionHeuristic>
-    <queuedEntityPlacer>
-      <entitySelector id="placerEntitySelector">
-        <entityClass>examination.domain.Exam</entityClass>
-        <cacheType>PHASE</cacheType>
-        <selectionOrder>SORTED</selectionOrder>
-        <sorterManner>DECREASING_DIFFICULTY</sorterManner>
-      </entitySelector>
-      <cartesianProductMoveSelector>
-        <changeMoveSelector>
-          <entitySelector mimicSelectorRef="placerEntitySelector"/>
-          <valueSelector>
-            <downcastEntityClass>examination.domain.LeadingExam</downcastEntityClass>
-            <variableName>period</variableName>
-            <cacheType>PHASE</cacheType>
-            <!--<selectionOrder>SORTED</selectionOrder>-->
-            <!--<sorterManner>INCREASING_STRENGTH</sorterManner>-->
-          </valueSelector>
-        </changeMoveSelector>
-        <changeMoveSelector>
-          <entitySelector mimicSelectorRef="placerEntitySelector"/>
-          <valueSelector>
-            <variableName>room</variableName>
-            <cacheType>PHASE</cacheType>
-            <selectionOrder>SORTED</selectionOrder>
-            <sorterManner>INCREASING_STRENGTH</sorterManner>
-          </valueSelector>
-        </changeMoveSelector>
-      </cartesianProductMoveSelector>
-    </queuedEntityPlacer>
-  </constructionHeuristic>
-  <localSearch>
-    <unionMoveSelector>
-      <cartesianProductMoveSelector>
-        <changeMoveSelector>
-          <entitySelector id="cartesianProductEntitySelector">
-            <entityClass>examination.domain.Exam</entityClass>
-          </entitySelector>
-          <valueSelector>
-            <variableName>room</variableName>
-          </valueSelector>
-        </changeMoveSelector>
-        <changeMoveSelector>
-          <entitySelector mimicSelectorRef="cartesianProductEntitySelector"/>
-          <valueSelector>
-            <downcastEntityClass>examination.domain.LeadingExam</downcastEntityClass>
-            <variableName>period</variableName>
-          </valueSelector>
-        </changeMoveSelector>
-      </cartesianProductMoveSelector>
-      <swapMoveSelector>
-        <entitySelector>
-          <entityClass>examination.domain.LeadingExam</entityClass>
-        </entitySelector>
-      </swapMoveSelector>
-    </unionMoveSelector>
-    <acceptor>
-      <entityTabuSize>10</entityTabuSize>
-    </acceptor>
-    <forager>
-      <acceptedCountLimit>2000</acceptedCountLimit>
-    </forager>
-  </localSearch>
 </solver>
