+++ projeto/calendar_generator_engine/src/main/java/examination/domain/Auxiliar.java
@@ -3,13 +3,11 @@ package examination.domain;
 /**
  * @author Duarte
  * @version 1.0
- * @created 18-fev-2016 16:42:19
+ * @created 18-fev-2016 16:42:18
  */
-public class Student {
+public class Auxiliar {
 
-	private long id;
-
-	public Student(){
+	public Auxiliar(){
 
 	}
