+++ projeto/calendar_generator_engine/src/main/java/examination/domain/Exam.java
@@ -1,115 +1,28 @@
-/*
- * Copyright 2010 JBoss Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package examination.domain;
 
-import com.thoughtworks.xstream.annotations.XStreamInclude;
-import org.optaplanner.core.api.domain.entity.PlanningEntity;
-import org.optaplanner.core.api.domain.variable.PlanningVariable;
-import common.domain.AbstractPersistable;
-import examination.domain.solver.ExamDifficultyWeightFactory;
-import examination.domain.solver.RoomStrengthWeightFactory;
+import java.util.List;
 
-@PlanningEntity(difficultyWeightFactoryClass = ExamDifficultyWeightFactory.class)
-@XStreamInclude({
-        LeadingExam.class,
-        FollowingExam.class
-})
-public abstract class Exam extends AbstractPersistable {
+/**
+ * @author Duarte
+ * @version 1.0
+ * @created 18-fev-2016 16:42:18
+ */
+public class Exam {
 
-    protected Topic topic;
+	private int numStudents;
+	private boolean pc;
+	/**
+	 * shadow variable
+	 */
+	private List<RoomPeriod> roomPeriods;
+	public Topic m_Topic;
 
-    // Planning variables: changes during planning, between score calculations.
-    protected Room room;
+	public Exam(){
 
-    public Topic getTopic() {
-        return topic;
-    }
+	}
 
-    public void setTopic(Topic topic) {
-        this.topic = topic;
-    }
+	public void finalize() throws Throwable {
 
-    @PlanningVariable(valueRangeProviderRefs = {"roomRange"},
-            strengthWeightFactoryClass = RoomStrengthWeightFactory.class)
-    public Room getRoom() {
-        return room;
-    }
-
-    public void setRoom(Room room) {
-        this.room = room;
-    }
-
-    // ************************************************************************
-    // Complex methods
-    // ************************************************************************
-
-    public abstract Period getPeriod();
-
-    public int getTopicDuration() {
-        return getTopic().getDuration();
-    }
-
-    public int getTopicStudentSize() {
-        return getTopic().getStudentSize();
-    }
-
-    public int getDayIndex() {
-        Period period = getPeriod();
-        if (period == null) {
-            return Integer.MIN_VALUE;
-        }
-        return period.getDayIndex();
-    }
-
-    public int getPeriodIndex() {
-        Period period = getPeriod();
-        if (period == null) {
-            return Integer.MIN_VALUE;
-        }
-        return period.getPeriodIndex();
-    }
-
-    public int getPeriodDuration() {
-        Period period = getPeriod();
-        if (period == null) {
-            return Integer.MIN_VALUE;
-        }
-        return period.getDuration();
-    }
-
-    public boolean isTopicFrontLoadLarge() {
-        return topic.isFrontLoadLarge();
-    }
-
-    public boolean isPeriodFrontLoadLast() {
-        Period period = getPeriod();
-        if (period == null) {
-            return false;
-        }
-        return period.isFrontLoadLast();
-    }
-
-    public String getLabel() {
-        return Long.toString(topic.getId());
-    }
-
-    @Override
-    public String toString() {
-        return topic.toString();
-    }
+	}
 
 }
