+++ projeto/calendar_generator_engine/src/main/java/examination/domain/Examination.java
@@ -1,171 +1,65 @@
-/*
- * Copyright 2010 JBoss Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package examination.domain;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import com.thoughtworks.xstream.annotations.XStreamAlias;
-import com.thoughtworks.xstream.annotations.XStreamConverter;
 import org.optaplanner.core.api.domain.solution.PlanningEntityCollectionProperty;
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.api.domain.solution.Solution;
 import org.optaplanner.core.api.domain.valuerange.ValueRangeProvider;
 import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore;
-import org.optaplanner.core.impl.score.buildin.hardsoft.HardSoftScoreDefinition;
-import common.domain.AbstractPersistable;
-import examination.domain.solver.TopicConflict;
-import org.optaplanner.persistence.xstream.impl.score.XStreamScoreConverter;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
+/**
+ * @author Gustavo
+ * @version 1.0
+ * @created 18-fev-2016 16:42:18
+ */
 @PlanningSolution()
-@XStreamAlias("Examination")
-public class Examination extends AbstractPersistable
-        implements Solution<HardSoftScore> {
+public class Examination implements Solution<HardSoftScore> {
+	private List<RoomPeriod> roomPeriodList;
 
-    private InstitutionParametrization institutionParametrization;
+	public List<Exam> examList;
+	public InstitutionParametrization m_InstitutionParametrization;
 
-    private List<Student> studentList;
-    private List<Topic> topicList;
-    private List<Period> periodList;
-    private List<Room> roomList;
+	public Examination(){
+		this.examList = new ArrayList<Exam>();
+		this.roomPeriodList = new ArrayList<RoomPeriod>();
+	}
 
-    private List<PeriodPenalty> periodPenaltyList;
-    private List<RoomPenalty> roomPenaltyList;
+	@PlanningEntityCollectionProperty
+	public List<RoomPeriod> getRoomPeriodList() {
+		return roomPeriodList;
+	}
+	public void addRoomPeriod(RoomPeriod roomPeriod){ this.roomPeriodList.add(roomPeriod); }
 
-    private List<Exam> examList;
+	public void finalize() throws Throwable {
 
-    @XStreamConverter(value = XStreamScoreConverter.class, types = {HardSoftScoreDefinition.class})
-    private HardSoftScore score;
+	}
 
-    public InstitutionParametrization getInstitutionParametrization() {
-        return institutionParametrization;
-    }
+	@ValueRangeProvider(id = "examRange")
+	public List<Exam> getExamList() {
+		return this.examList;
+	}
 
-    public void setInstitutionParametrization(InstitutionParametrization institutionParametrization) {
-        this.institutionParametrization = institutionParametrization;
-    }
+	public void setExamList(List<Exam> periodList) {
+		this.examList = periodList;
+	}
 
-    public List<Student> getStudentList() {
-        return studentList;
-    }
+	@Override
+	public HardSoftScore getScore() {
+		return HardSoftScore.valueOf(0,0);
+	}
 
-    public void setStudentList(List<Student> studentList) {
-        this.studentList = studentList;
-    }
+	@Override
+	public void setScore(HardSoftScore score) {
 
-    public List<Topic> getTopicList() {
-        return topicList;
-    }
+	}
 
-    public void setTopicList(List<Topic> topicList) {
-        this.topicList = topicList;
-    }
-
-    @ValueRangeProvider(id = "periodRange")
-    public List<Period> getPeriodList() {
-        return periodList;
-    }
-
-    public void setPeriodList(List<Period> periodList) {
-        this.periodList = periodList;
-    }
-
-    @ValueRangeProvider(id = "roomRange")
-    public List<Room> getRoomList() {
-        return roomList;
-    }
-
-    public void setRoomList(List<Room> roomList) {
-        this.roomList = roomList;
-    }
-
-    public List<PeriodPenalty> getPeriodPenaltyList() {
-        return periodPenaltyList;
-    }
-
-    public void setPeriodPenaltyList(List<PeriodPenalty> periodPenaltyList) {
-        this.periodPenaltyList = periodPenaltyList;
-    }
-
-    public List<RoomPenalty> getRoomPenaltyList() {
-        return roomPenaltyList;
-    }
-
-    public void setRoomPenaltyList(List<RoomPenalty> roomPenaltyList) {
-        this.roomPenaltyList = roomPenaltyList;
-    }
-
-    @PlanningEntityCollectionProperty
-    public List<Exam> getExamList() {
-        return examList;
-    }
-
-    public void setExamList(List<Exam> examList) {
-        this.examList = examList;
-    }
-
-    public HardSoftScore getScore() {
-        return score;
-    }
-
-    public void setScore(HardSoftScore score) {
-        this.score = score;
-    }
-
-    // ************************************************************************
-    // Complex methods
-    // ************************************************************************
-
-    public Collection<? extends Object> getProblemFacts() {
-        List<Object> facts = new ArrayList<Object>();
-        facts.add(institutionParametrization);
-        // Student isn't used in the DRL at the moment
-        // Notice that asserting them is not a noticable performance cost, only a memory cost.
-        // facts.addAll(studentList);
-        facts.addAll(topicList);
-        facts.addAll(periodList);
-        facts.addAll(roomList);
-        facts.addAll(periodPenaltyList);
-        facts.addAll(roomPenaltyList);
-        // A faster alternative to a insertLogicalTopicConflicts rule.
-        facts.addAll(precalculateTopicConflictList());
-        // Do not add the planning entity's (examList) because that will be done automatically
-        return facts;
-    }
-
-    private List<TopicConflict> precalculateTopicConflictList() {
-        List<TopicConflict> topicConflictList = new ArrayList<TopicConflict>();
-        for (Topic leftTopic : topicList) {
-            for (Topic rightTopic : topicList) {
-                if (leftTopic.getId() < rightTopic.getId()) {
-                    int studentSize = 0;
-                    for (Student student : leftTopic.getStudentList()) {
-                        if (rightTopic.getStudentList().contains(student)) {
-                            studentSize++;
-                        }
-                    }
-                    if (studentSize > 0) {
-                        topicConflictList.add(new TopicConflict(leftTopic, rightTopic, studentSize));
-                    }
-                }
-            }
-        }
-        return topicConflictList;
-    }
-
+	@Override
+	public Collection<?> getProblemFacts() {
+		List<Object> facts = new ArrayList<Object>();
+		facts.addAll(examList);
+		return facts;
+	}
 }
