+++ projeto/calendar_generator_engine/src/main/java/examination/domain/PeriodTime.java
@@ -1,0 +1,10 @@
+package examination.domain;
+
+/**
+ * Created by Duarte on 18/02/2016.
+ */
+public enum PeriodTime {
+    NINE_AM,
+    ONE_PM,
+    FIVE_PM
+}
