+++ projeto/calendar_generator_engine/src/main/java/examination/domain/Room.java
@@ -1,53 +1,47 @@
-/*
- * Copyright 2010 JBoss Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package examination.domain;
 
-import com.thoughtworks.xstream.annotations.XStreamAlias;
-import common.domain.AbstractPersistable;
+/**
+ * @author Duarte
+ * @version 1.0
+ * @created 18-fev-2016 16:42:19
+ */
+public class Room {
 
-@XStreamAlias("Room")
-public class Room extends AbstractPersistable  {
+	private int capacity;
+	private String codRoom;
+	private boolean pc;
 
-    private int capacity;
-    private int penalty;
+	public Room(String codRoom, int capacity, boolean pc){
+		this.capacity = capacity;
+		this.codRoom = codRoom;
+		this.pc = pc;
+	}
 
-    public int getCapacity() {
-        return capacity;
-    }
+	public void finalize() throws Throwable {
 
-    public void setCapacity(int capacity) {
-        this.capacity = capacity;
-    }
+	}
 
-    public int getPenalty() {
-        return penalty;
-    }
+	public String getCodRoom() {
+		return codRoom;
+	}
 
-    public void setPenalty(int penalty) {
-        this.penalty = penalty;
-    }
+	public void setCodRoom(String codRoom) {
+		this.codRoom = codRoom;
+	}
 
-    public String getLabel() {
-        return Long.toString(id);
-    }
+	public int getCapacity() {
+		return capacity;
+	}
 
-    @Override
-    public String toString() {
-        return Long.toString(id);
-    }
+	public void setCapacity(int capacity) {
+		this.capacity = capacity;
+	}
 
+	public boolean isPc() {
+		return pc;
+	}
+
+	public void setPc(boolean pc) {
+		this.pc = pc;
+	}
 }
