+++ projeto/calendar_generator_engine/src/main/java/examination/example_domain/RoomPenalty.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2010 JBoss Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package examination.example_domain;
+
+import com.thoughtworks.xstream.annotations.XStreamAlias;
+import common.domain.AbstractPersistable;
+
+@XStreamAlias("RoomPenalty")
+public class RoomPenalty extends AbstractPersistable {
+
+    private RoomPenaltyType roomPenaltyType;
+    private Topic topic;
+
+    public RoomPenaltyType getRoomPenaltyType() {
+        return roomPenaltyType;
+    }
+
+    public void setRoomPenaltyType(RoomPenaltyType roomPenaltyType) {
+        this.roomPenaltyType = roomPenaltyType;
+    }
+
+    public Topic getTopic() {
+        return topic;
+    }
+
+    public void setTopic(Topic topic) {
+        this.topic = topic;
+    }
+
+    @Override
+    public String toString() {
+        return roomPenaltyType + "@" + topic.getId();
+    }
+
+}
