+++ projeto/calendar_generator_engine/src/main/java/examination/example_domain/solver/ExamDifficultyWeightFactory.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package examination.domain.solver;
+package examination.example_domain.solver;
 
-import examination.domain.PeriodPenalty;
+import examination.example_domain.PeriodPenalty;
 import org.apache.commons.lang3.builder.CompareToBuilder;
 import org.optaplanner.core.impl.heuristic.selector.common.decorator.SelectionSorterWeightFactory;
-import examination.domain.Exam;
-import examination.domain.Examination;
-import examination.domain.LeadingExam;
+import examination.example_domain.Exam;
+import examination.example_domain.Examination;
+import examination.example_domain.LeadingExam;
 
 
 public class ExamDifficultyWeightFactory implements SelectionSorterWeightFactory<Examination, Exam> {
